refactor(feedback): split body formatter out of github_issue_reporter (Refs #563)
- github_issue_reporter.dart: 352 -> ~198 LOC (HTTP + rate-limit shell only).
- lib/core/feedback/github_issue_body_formatter.dart (new): issue body construction (Markdown table, sanitization, scan-kind label, EXIF strip).
- GithubIssueReporter public API (reportBadScan, ScanKind, GithubReporterException) unchanged; tests stayed green.
